Renames the tests to use a fixed buffer.
